remove duplicated 2021/Q1 data
I had some API errors (temporary outage or network problems) while performing the update script, so I retried it several times and it seemed to work. However, the retry has caused the script to add the 2021/Q1 data multiple times to some of the datasets. This commits removes the duplicated data from the dataset.